MyTableとMyTableUnitに分割
