+++ src/components/sampleStories/actions.stories.tsx
@@ -1,11 +1,11 @@
-import React from "react";
-import { action } from "@storybook/addon-actions";
-import "semantic-ui-css/semantic.min.css";
-import { Button } from "semantic-ui-react";
+import React from 'react';
+import { action } from '@storybook/addon-actions';
+import 'semantic-ui-css/semantic.min.css';
+import { Button } from 'semantic-ui-react';
 
 export default {
   component: Button,
-  title: "actionButton",
+  title: 'actionButton',
 };
 
-export const button = () => <Button onClick={action("clicked")}>Click Here</Button>;
+export const button = () => <Button onClick={action('clicked')}>Click Here</Button>;
