+++ src/components/table/tableUnit.stories.tsx
@@ -3,7 +3,7 @@ import 'semantic-ui-css/semantic.min.css';
 import { Provider } from 'react-redux';
 import { Table } from 'semantic-ui-react';
 import { store } from '../../stores/index';
-import { MyTableUnit } from './tableUnit';
+import { MyTableUnit } from './myTableUnit';
 
 export default {
   component: MyTableUnit,
